product mapping test + changing name of variables to user in user repository

+++ flutter_project/lib/models/product.dart
@@ -32,7 +32,7 @@ class Product {
       id: json['id'],
       title: json['title'],
       description: json['description'],
-      price: json['price'],
+      price: ((json['price']) as int).toDouble(),
       discountPercentage: json['discountPercentage'],
       rating: json['rating'],
       stock: json['stock'],
